refactor(graphics): extract camera math helpers (Issue #284 Phase1)
