+++ viewmodel/graphics/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod camera;
+mod camera_math;
 
 pub use camera::{build_view_projection_matrix, Camera, CameraControlSensitivity};
